requirements.txt
mysql-connector-python
PyYAML
scikit-learn
nltk
newspaper3k
url-parser
searchtweets-v2
url-parser
tweepy